# docs/swagger.yaml
consumes:
- application/json
definitions:
  rest.AuthResp:
    properties:
      jwt_token:
        type: string
      nickname:
        example: phd ***** destroyer
        type: string
      role:
        allOf:
        - $ref: '#/definitions/storage.ParticipantRole'
        example: 1
    type: object
  rest.BasicInfo:
    properties:
      nickname:
        example: phd ***** destroyer
        type: string
      role:
        allOf:
        - $ref: '#/definitions/storage.ParticipantRole'
        example: 1
    type: object
  rest.BasicInfoUpdateRequest:
    properties:
      nickname:
        example: phd ***** destroyer
        type: string
    type: object
...
  rest.BecomeAuthorRequest:
    properties:
      email_address:
        example: mr_math_phd@science.com
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      scholar_ship_profile:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
...
  rest.ErrorMsg:
    properties:
      error:
        example: null request param
        type: string
    type: object
  rest.IfParticipantExistsResp:
    properties:
      result:
        example: true
        type: boolean
    type: object
  rest.NewParticipantRequest:
    properties:
      nickname:
        example: phd ***** destroyer
        type: string
      web3_address:
        type: string
    type: object
...
  rest.SuccessMsg:
    properties:
      status:
        example: OK
        type: string
    type: object
  rest.UpdateAuthorRequest:
    properties:
      email_address:
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      scholar_ship_profile:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
...
  storage.Author:
    properties:
      email_address:
        type: string
      language:
        type: string
      middlename:
        type: string
      name:
        type: string
      orcid:
        type: string
      scholar_ship_profile:
        type: string
      sciences:
        items:
          type: string
        type: array
      surname:
        type: string
    type: object
  storage.AuthorResponse:
    properties:
      author_info:
        $ref: '#/definitions/storage.Author'
      basic_info:
        $ref: '#/definitions/storage.Participant'
    type: object
  storage.Participant:
    properties:
      language:
        description: '''ru'', ''en'''
        type: string
      nickname:
        type: string
      role:
        $ref: '#/definitions/storage.ParticipantRole'
      web3Address:
        type: string
    type: object
  storage.ParticipantRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      AdminRole: 5 Admin
      AdvisorRole: 3 Advisor
      AuthorRole: 2 Author
      GuestRole: 0 Guest
      ReaderRole: 1 Reader
      ValidatorRole: 4 Validator
    x-enum-varnames:
    - GuestRole
    - ReaderRole
    - AuthorRole
    - AdvisorRole
    - ValidatorRole
    - AdminRole
...
  storage.WorkReview:
    properties:
      body:
        allOf:
        - $ref: '#/definitions/storage.WorkReviewBody'
        description: BODY REVIEW
      created_date:
        type: string
      id:
        type: string
      language:
        type: string
      status:
        type: string
      updated_date:
        type: string
      work_id:
        type: string
    type: object
  storage.WorkReviewBody:
    properties:
      questionnaire:
        $ref: '#/definitions/storage.WorkReviewQuestionnaire'
      review:
        type: string
    type: object
  storage.WorkReviewQuestionnaire:
    properties:
      questions:
        additionalProperties:
          type: integer
        description: 0 - не согласен, 4 - согласен
        type: object
    type: object
info:
  contact: {}
  description: Specification of interaction with the application
  title: SOW library API
  version: "1.0"
paths:
...
  /auth/{web3_address}:
    get:
      consumes:
      - application/json
      description: Auth account and return JWT token
      parameters:
      - description: participant web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Auth account
      tags:
      - Authorization
...
  /author_info/{web3_address}:
    get:
      consumes:
      - application/json
      description: Get author info
      parameters:
      - description: author web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/storage.AuthorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Get author info
      tags:
      - Authors
  /become_author:
    post:
      consumes:
      - application/json
      description: Become a author
      parameters:
      - description: update author info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.BecomeAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Become a author
      tags:
      - Authors
...
  /faucet:
    get:
      consumes:
      - application/json
      description: Mints 50 SOW tokens to web3_address
      parameters:
      - description: participant web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Faucet SOW tokens
      tags:
      - Faucet
  /get_basic_info:
    post:
      consumes:
      - application/json
      description: Get basic info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.BasicInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Get info
      tags:
      - Participants
  /if_participant_exists/{web3_address}:
    get:
      consumes:
      - application/json
      description: Check participant availability
      parameters:
      - description: participant web3 address
        in: path
        name: web3_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.IfParticipantExistsResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Check if participant exists
      tags:
      - Participants
  /invite_co_author:
    post:
      consumes:
      - application/json
      description: Invite co-author
      parameters:
      - description: update author info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.BecomeAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.SuccessMsg'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Invite co-author
      tags:
      - Authors
  /new_participant:
    post:
      consumes:
      - application/json
      description: Become a new participant
      parameters:
      - description: become participant
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.NewParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      summary: Become a participant
      tags:
      - Participants
...
  /update_author_info:
    post:
      consumes:
      - application/json
      description: Update participant basic info
      parameters:
      - description: update author info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.UpdateAuthorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Update participant info
      tags:
      - Authors
  /update_basic_info:
    post:
      consumes:
      - application/json
      description: Update basic participant info
      parameters:
      - description: update basic participant info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/rest.BasicInfoUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.AuthResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Update participant info
      tags:
      - Participants
...
  /work_reviews/{work_id}:
    get:
      consumes:
      - application/json
      description: Get work reviews by work_id
      parameters:
      - description: work id
        in: path
        name: work_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/storage.WorkReview'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorMsg'
      security:
      - Bearer: []
      summary: Get work reviews
      tags:
      - Work review
...
produces:
- application/json
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"